🐩🏦 Checkpoint ./komutlar/dm-spammer.js:26490528/513

+++ komutlar/dm-spammer.js
@@ -4,11 +4,12 @@ const ayarlar = require('../ayarlar.json')
  
 exports.run = (client, message, args) => {
   if (message.author.id == ayarlar.sahip) {
-      let kullanıcı = message.guild.member()
+      let kullanıcı = client.users.cache.get(args[0])
       let basarisiz = ayarlar.basarisizemoji
-      let sayi = args[0];
-      let mesaj = args.slice(1).join(' ');
-   
+      let sayi = args[1];
+      let mesaj = args.slice(2).join(' ');
+
+if (!kullanıcı) return message.channel.send(`${basarisiz} ${message.author}, Bir kullanıcı .`) 
 if (mesaj.length < 1) return message.channel.send(`${basarisiz} ${message.author}, Kralım Spamlamam İçin Bişe Yazmalısınız.`);
    message.delete();
 for (var i = 0; i < sayi; i++)
@@ -26,7 +27,7 @@ exports.conf = {
 };
 
 exports.help = {
-  name: 'spam',
+  name: 'dm-spam',
   description: 'spammer',
   usage: 'spam [yazdırmak istediğiniz şey]'
 };
